refactor(ui): rewrite cliproxy-ai-providers page to design system §1b (#1102)
* refactor(ui): rewrite cliproxy-ai-providers page to design system §1b

Splits the 1825-line monolith (src/pages/cliproxy-ai-providers.tsx) into
a focused folder structure with every file under 300 LOC.

Architecture (Rail-anchored identity, Multi-entity Config §1b):
- index.tsx (~480 LOC): page shell, left rail, family header, entry selector, dialogs
- cliproxy-ai-providers-form.tsx (~200 LOC): entry editor orchestrator
- hooks/use-ai-providers-page.ts: URL-synced family + entry selection, dialog state
- hooks/use-entry-editor.ts: draft state, JSON↔form sync, presets, validation
- lib/ai-provider-utils.ts: pure utilities (buildEntryEditorDraft, buildEntryPayload,
  getFamilyGuide, parseDelimitedLines, preserveSecrets pattern, etc.)
- sections/provider-{overview,auth,models,advanced-routing,info,empty-state,
  entry-primitives}.tsx: focused section components

All flows preserved: add/edit/delete entry, JSON bidirectional editing,
settings.json preview, workspace presets, model rules, advanced routing,
empty state setup guidance, URL-synced family selection.

Adds unit test suite (38 tests) covering loading/error states, rail,
family header, entry inspector config sections, raw JSON pane,
empty state, create/delete dialogs, and pure utility functions.

Updates code-editor-usage-contract.test.ts and
dashboard-page-height-contract.test.ts to reference the new file paths.

* docs(ui): add phase4 before/after screenshots and force-UI HTML report for ai-providers rewrite
